일단 약관동의하고 홈은 넘어갔어
