+++ app/login.tsx
@@ -1,7 +1,3 @@
 // app/login.tsx
-import React from 'react';
 import LoginScreen from '../screens/LoginScreen';
-
-export default function LoginPage() {
-  return <LoginScreen/>;
-}
+export default LoginScreen;
