+++ app/signup.tsx
@@ -1,7 +1,3 @@
 // app/signup.tsx
-import React from 'react';
-import SignUpScreen from '../screens/SignUpScreen'; // 경로는 프로젝트 구조에 맞게 조정
-
-export default function SignupPage() {
-  return <SignUpScreen />;
-}
+import SignUpNicknameScreen from '../screens/SignUpNicknameScreen';
+export default SignUpNicknameScreen;
